+++ src/Entry.cpp
@@ -1,7 +1,5 @@
 #include "Entry.h"
 
-#include <ila/event/minecraft/world/ExplosionEvent.h>
-
 #include <ll/api/Config.h>
 #include <ll/api/event/EventBus.h>
 #include <ll/api/memory/Hook.h>
@@ -9,15 +7,15 @@
 #include <ll/api/service/Bedrock.h>
 
 #include <mc/deps/core/math/Vec3.h>
-#include <mc/legacy/ActorUniqueID.h> // IWYU pragma: keep
 #include <mc/world/actor/player/Player.h>
 #include <mc/world/level/BlockPos.h>
+#include <mc/world/level/BlockSource.h>
+#include <mc/world/level/Explosion.h>
 #include <mc/world/level/Level.h>
 #include <mc/world/level/block/BedBlock.h>
 #include <mc/world/level/block/Block.h>
 #include <mc/world/level/block/RespawnAnchorBlock.h>
 #include <mc/world/level/dimension/Dimension.h>
-
 
 namespace BanExplosion {
 
@@ -27,16 +25,13 @@ LL_TYPE_INSTANCE_HOOK( // NOLINT
     BedBlockUseHook,
     HookPriority::Normal,
     BedBlock,
-    &BedBlock::$use,
-    bool,
-    Player&         player,
-    BlockPos const& pos,
-    uchar           face
+    &BedBlock::use,
+    void,
+    BlockEvents::BlockPlayerInteractEvent& eventData
 ) {
-    mTypeName         = getTypeName();
-    const auto result = origin(player, pos, face);
-    mTypeName         = "";
-    return result;
+    mTypeName = getTypeName();
+    origin(eventData);
+    mTypeName = "";
 }
 
 LL_TYPE_STATIC_HOOK( // NOLINT
@@ -55,6 +50,35 @@ LL_TYPE_STATIC_HOOK( // NOLINT
     mTypeName = "";
 }
 
+LL_TYPE_INSTANCE_HOOK( // NOLINT
+    ExplodeHook, HookPriority::Normal, Explosion, &Explosion::explode, bool) {
+    auto& typeName = mTypeName;
+    if (typeName.empty()) {
+        const auto* actor = ll::service::getLevel()->fetchEntity(mSourceID, false);
+        typeName          = actor == nullptr ? mRegion.getBlock(mPos.get()).getTypeName() : actor->getTypeName();
+    }
+    const auto& config = Entry::getInstance().getConfig();
+    const auto& logger = Entry::getInstance().getSelf().getLogger();
+    const auto& setting =
+        config.explosionSetting.contains(typeName) ? config.explosionSetting.at(typeName) : config.defaultSetting;
+    if (logger.getLevel() >= ll::io::LogLevel::Debug) {
+        logger.debug(
+            "Explosion: {0} at {1}(DimId: {2}), setting: {3}",
+            typeName,
+            mPos->toString(),
+            mRegion.getDimension().mName.get(),
+            ll::reflection::serialize<nlohmann::ordered_json>(setting).value().dump(4)
+        );
+    }
+    if (!setting.allowExplosion) return false;
+    if (!setting.allowDestroy) mBreaking = false;
+    if (!setting.allowFire) mFire = false;
+    if (setting.maxRadius < mRadius) mRadius = setting.maxRadius;
+    return origin();
+}
+
+using Hooks = ll::memory::HookRegistrar<BedBlockUseHook, RespawnAnchorBlockExplodeHook, ExplodeHook>;
+
 Entry& Entry::getInstance() {
     static Entry instance;
     return instance;
@@ -69,40 +93,13 @@ bool Entry::load() {
     return true;
 }
 
-bool Entry::enable() {
-    mListener = ll::event::EventBus::getInstance().emplaceListener<ila::mc::ExplosionBeforeEvent>(
-        [this](ila::mc::ExplosionBeforeEvent& event) -> void {
-            auto& explosion = event.explosion();
-            auto  typeName  = mTypeName;
-            if (typeName.empty()) {
-                const auto* actor = ll::service::getLevel()->fetchEntity(explosion.mSourceID, false);
-                typeName          = actor == nullptr ? event.blockSource().getBlock(explosion.mPos.get()).getTypeName()
-                                                     : actor->getTypeName();
-            }
-            auto setting = getConfig().explosionSetting.contains(typeName) ? getConfig().explosionSetting[typeName]
-                                                                           : getConfig().defaultSetting;
-            if (getSelf().getLogger().getLevel() >= ll::io::LogLevel::Debug) {
-                getSelf().getLogger().debug(
-                    "Explosion: {0} at {1}(DimId: {2}), setting: {3}",
-                    typeName,
-                    explosion.mPos->toString(),
-                    explosion.mRegion.getDimension().mName.get(),
-                    ll::reflection::serialize<nlohmann::ordered_json>(setting).value().dump(4)
-                );
-            }
-            if (!setting.allowExplosion) return event.cancel();
-            if (!setting.allowDestroy) explosion.mBreaking = false;
-            if (!setting.allowFire) explosion.mFire = false;
-            if (setting.maxRadius < explosion.mRadius) explosion.mRadius = setting.maxRadius;
-        }
-    );
-    ll::memory::HookRegistrar<BedBlockUseHook, RespawnAnchorBlockExplodeHook>::hook();
+bool Entry::enable() /* NOLINT */ {
+    Hooks::hook();
     return true;
 }
 
 bool Entry::disable() /* NOLINT */ {
-    ll::event::EventBus::getInstance().removeListener(mListener);
-    ll::memory::HookRegistrar<BedBlockUseHook, RespawnAnchorBlockExplodeHook>::unhook();
+    Hooks::unhook();
     return true;
 }